move QueueProcessor to the async package
